src/main/java/ru/netology/films/MovieManager.java
package ru.netology.films;

public class MovieManager {
    private String[] movies = new String[0];
    private int limit;

    public MovieManager() {
        this.limit = 5;
    }

    public MovieManager(int limit) {
        this.limit = limit;
    }

    public void add(String film) {
        String[] tmp = new String[movies.length + 1];
        for (int i = 0; i < movies.length; i++) {
            tmp[i] = movies[i];
        }
        tmp[tmp.length - 1] = film;
        movies = tmp;
    }

    public String[] findall() {
        return movies;
    }

    public String[] findLast() {
        int totalFilms;

        if (movies.length < limit) {
            totalFilms = movies.length;
        } else {
            totalFilms = limit;
        }

        String[] tmp = new String[totalFilms];
        for (int i = 0; i < tmp.length; i++) {
            tmp[i] = movies[movies.length - 1 - i];
        }
        return tmp;
    }
}

